# Conflicts:
#	app/src/main/java/com/it/mahaalrasheed/route/map.java

+++ app/src/main/java/com/it/mahaalrasheed/route/FavoriteClass.java
@@ -8,16 +8,17 @@ import io.realm.annotations.PrimaryKey;
  */
 public class FavoriteClass extends RealmObject {
     @PrimaryKey
-    private long id ;
+    private int id ;
+
     private double lng;
     private double lat;
     private String name;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
